tests/requirements.txt
pytest>=7.0
python-docx>=1.1.0
pydantic>=2.0
pydantic-settings>=2.0
httpx>=0.25
